moved Sum,Mex,Mey & Met definitions to cmsl1t.energySums
