+++ cmsl1t/energySums.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+from collections import namedtuple
+
+EnergySum = namedtuple('Sum', ['et'])
+Mex = namedtuple('Mex', ['ex'])
+Mey = namedtuple('Mey', ['ey'])
+Met = namedtuple('Met', ['et', 'phi'])
